net-analyzer/wireshark: Remove /usr/share/appdata when present
Package-Manager: Portage-2.3.99, Repoman-2.3.22
Fixes: b0222a2948853f723eb7b52e0b972702f1684b29

+++ net-analyzer/wireshark/wireshark-3.2.4.ebuild
@@ -226,7 +226,9 @@ src_install() {
 		done
 	fi
 
-	rm -r "${D}"/usr/share/appdata || die
+	if [[ -d "${D}"/usr/share/appdata ]]; then
+		rm -r "${D}"/usr/share/appdata || die
+	fi
 }
 
 pkg_postinst() {
