+++ net-analyzer/wireshark/wireshark-99999999.ebuild
@@ -220,7 +220,9 @@ src_install() {
 		done
 	fi
 
-	rm -r "${D}"/usr/share/appdata || die
+	if [[ -d "${D}"/usr/share/appdata ]]; then
+		rm -r "${D}"/usr/share/appdata || die
+	fi
 }
 
 pkg_postinst() {
